+++ apiserver/server_errors.go
@@ -1,4 +1,4 @@
-package eventserver
+package apiserver
 
 import (
 	"fmt"
