+++ apiserver/server_plans_handler.go
@@ -1,4 +1,4 @@
-package eventserver
+package apiserver
 
 import (
 	"net/http"
