+++ apiserver/server_plans_handler_test.go
@@ -1,4 +1,4 @@
-package eventserver_test
+package apiserver_test
 
 import (
 	"context"
@@ -11,7 +11,7 @@ import (
 	"github.com/alphagov/paas-billing/fakes"
 	"github.com/labstack/echo"
 
-	. "github.com/alphagov/paas-billing/eventserver"
+	. "github.com/alphagov/paas-billing/apiserver"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
